Attendants line up at attractions and enter attraction
Attendants now line up at enter attractions

Next on agenda:
- Have attendants leave attraction and choose new location
- Make sure attendants only enter attraction when they have actually arrived in the queue.

+++ Assets/_Scripts/_Scriptable Objects/Events/Attraction.cs
@@ -6,7 +6,7 @@ public class Attraction : Event
     [Header("Attraction Attributes")]
     public int attendentsPerRide;
     public int lineCapaticy;
-    public float rideLength;
+    public int rideLength;
 
     [Header("Express Attributes")]
     public bool usesExpressSystem;
